+++ Skin/Retail/SharedXML/AudioOptionsPanels.lua
@@ -68,22 +68,4 @@ function private.FrameXML.AudioOptionsPanels()
     Skin.OptionsSliderTemplate(_G.AudioOptionsSoundPanelMusicVolume)
     Skin.OptionsSliderTemplate(_G.AudioOptionsSoundPanelAmbienceVolume)
     Skin.OptionsSliderTemplate(_G.AudioOptionsSoundPanelDialogVolume)
-
-    -----------------
-    -- Voice Panel --
-    -----------------
-    local AudioOptionsVoicePanel = _G.AudioOptionsVoicePanel
-    Skin.OptionsDropdownTemplate(AudioOptionsVoicePanel.OutputDeviceDropdown)
-    Skin.OptionsSliderTemplate(AudioOptionsVoicePanel.VoiceChatVolume)
-    Skin.OptionsSliderTemplate(AudioOptionsVoicePanel.VoiceChatDucking)
-    Skin.OptionsDropdownTemplate(AudioOptionsVoicePanel.MicDeviceDropdown)
-    Skin.OptionsSliderTemplate(AudioOptionsVoicePanel.VoiceChatMicVolume)
-    Skin.OptionsSliderTemplate(AudioOptionsVoicePanel.VoiceChatMicSensitivity)
-
-    local TestInputDevice = AudioOptionsVoicePanel.TestInputDevice
-    Skin.OptionsButtonTemplate(TestInputDevice.ToggleTest)
-    TestInputDevice.VUMeter:SetBackdrop(nil)
-    Skin.FrameTypeStatusBar(TestInputDevice.VUMeter.Status)
-
-    Skin.OptionsDropdownTemplate(AudioOptionsVoicePanel.ChatModeDropdown)
 end