+++ Skin/Retail/SharedXML/AudioOptionsPanelsVoice.lua
@@ -1,6 +1,5 @@
 local _, private = ...
 if private.isClassic then return end
-if private.isRetail then return end
 
 --[[ Lua Globals ]]
 -- luacheck: globals