+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
   * Error when loot skin is disabled
   * Chat bubble didn't work
+  * [retail] Volume sliders were white
   * [retail] Missed some textures on the spec frame
   * [retail] Display bug with quest detail
   * [retail] World quest header wasn't skinned